### 2024-02-19 #6379 @editorialbot
[PRE REVIEW]: FIGARO: hierarchical non-parametric inference for population studies
**Submitting author:** @sterinaldi (<a href="http://orcid.org/0000-0001-5799-4155">Stefano Rinaldi</a>)
**Repository:** https://github.com/sterinaldi/FIGARO
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.5.3
