### 2024-02-21 0c5cd50
added support for pdf and logpdf in absence of pre-instantiated mixture components
