### 2024-02-21 2c9f441
added make_comp flag to buiÃld_mixture
