### 2024-02-21 3fe9fca
mixtures without mn components (memory issue with large number of draws
