+++ figaro/marginal.py
@@ -51,8 +51,11 @@ def _marginalise(mix, axis = -1):
     means  = np.delete(mix.means, ax, axis = -1)
     covs   = np.delete(np.delete(mix.covs, ax, axis = -1), ax, axis = -2)
     bounds = np.delete(mix.bounds, ax, axis = 0)
-    
-    return mixture(means, covs, mix.w, bounds, dim, mix.n_cl, mix.n_pts, mix.alpha, probit = mix.probit)
+    if mix.components is None:
+        make_comp = False
+    else:
+        make_comp = True
+    return mixture(means, covs, mix.w, bounds, dim, mix.n_cl, mix.n_pts, mix.alpha, probit = mix.probit, make_comp = make_comp)
 
 def marginalise(draws, axis = -1):
     """
@@ -127,7 +130,11 @@ def _condition(mix, vals, dims, norm = True, filter = True, tol = 1e-4):
         idx_filt = [i in idx[m:] for i in range(len(ww))]
         if norm:
             log_weights -= logsumexp(log_weights[idx_filt])
-    return mixture(means[idx_filt], covs[idx_filt], np.exp(log_weights[idx_filt]), bounds, dim, len(log_weights[idx_filt]), mix.n_pts, mix.alpha, probit = mix.probit, log_w = log_weights[idx_filt])
+    if mix.components is None:
+        make_comp = False
+    else:
+        make_comp = True
+    return mixture(means[idx_filt], covs[idx_filt], np.exp(log_weights[idx_filt]), bounds, dim, len(log_weights[idx_filt]), mix.n_pts, mix.alpha, probit = mix.probit, log_w = log_weights[idx_filt], make_comp = make_comp)
 
 def condition(draws, vals, dims, norm = True, filter = True, tol = 1e-4):
     """
