### 2024-02-29 cde8fee
fixed bug while importing multiple gw parameters
