+++ figaro/_pipelines/par_probability_density.py
@@ -41,7 +41,7 @@ def main():
     parser.add_option("--ext", dest = "ext", type = "choice", choices = ['pkl', 'json'], help = "Format of mixture output file", default = 'json')
     parser.add_option("--inj_density", type = "string", dest = "inj_density_file", help = "Python module with injected density - please name the method 'density'", default = None)
     parser.add_option("--selfunc", type = "string", dest = "selfunc_file", help = "Python module with selection function - please name the method 'selection_function'", default = None)
-    parser.add_option("--parameter", type = "choice", dest = "par", help = "GW parameter(s) to be read from file", choices = supported_pars, default = None)
+    parser.add_option("--parameter", type = "string", dest = "par", help = "GW parameter(s) to be read from file", default = None)
     parser.add_option("--waveform", type = "choice", dest = "wf", help = "Waveform to load from samples file. To be used in combination with --parameter.", choices = ['combined', 'seob', 'imr'], default = 'combined')
     # Plot
     parser.add_option("-p", "--postprocess", dest = "postprocess", action = 'store_true', help = "Postprocessing", default = False)
@@ -94,7 +94,8 @@ def main():
     # Read parameter(s)
     if options.par is not None:
         options.par = options.par.split(',')
-
+        if not ([par in supported_pars for par in options.par]).all():
+            raise Exception("Please provide parameters from this list: "+', '.join(supported_pars[:-2]))
     # If provided, load injected density
     inj_density = None
     if options.inj_density_file is not None:
