### 2024-02-29 995306f
fixed bug while importing multiple gw parameters
