+++ figaro/_pipelines/probability_density.py
@@ -73,7 +73,7 @@ def main():
     # Read parameter(s)
     if options.par is not None:
         options.par = options.par.split(',')
-        if not ([par in supported_pars for par in options.par]).all():
+        if not np.all([par in supported_pars for par in options.par]):
             raise Exception("Please provide parameters from this list: "+', '.join(supported_pars[:-2]))
     # If provided, load injected density
     inj_density = None
