### 2024-03-15 16989b3
S gives nans
