### 2024-03-15 8085590
new_S problem
