### 2024-03-15 d9eb319
rescale-matrix nans
