### 2024-03-15 02001d4
rescale-matrix nans
