### 2024-03-15 624399f
prevented bug due to samples being exactly 0
