### 2024-03-19 aa0286e
added chi_p
