+++ figaro/load.py
@@ -23,6 +23,7 @@ GW_par = {'m1'                 : 'mass_1_source',
           'q'                  : 'mass_ratio',
           'eta'                : 'symmetric_mass_ratio',
           'chi_eff'            : 'chi_eff',
+          'chi_p'              : 'chi_p',
           'ra'                 : 'ra',
           'dec'                : 'dec',
           'luminosity_distance': 'luminosity_distance',
