### 2024-03-26 aba4c15
incorp selfunc
