+++ figaro/mixture.py
@@ -262,7 +262,7 @@ class _component_h:
     Returns:
         component_h: instance of component_h class
     """
-    def __init__(self, x, dim, prior, logL_D, mu_MC, sigma_MC):
+    def __init__(self, x, dim, prior, logL_D, mu_MC, sigma_MC, log_alpha_factor):
         self.dim    = dim
         self.N      = 1
         self.logL_D = logL_D
@@ -275,6 +275,7 @@ class _component_h:
         if dim == 1:
             self.mu = np.atleast_2d(self.mu).T
             self.sigma = np.atleast_2d(self.sigma).T
+        self.N_true = self.N/np.exp(log_alpha_factor[idx])
             
 class density:
     """
@@ -1138,22 +1139,24 @@ class HDPGMM(DPGMM):
     Child of DPGMM class
     
     Arguments:
-        iterable bounds:     boundaries of the rectangle over which the distribution is defined. It should be in the format [[xmin, xmax],[ymin, ymax],...]
-        iterable prior_pars: IW parameters
-        double alpha0:       initial guess for concentration parameter
-        double MC_draws:     number of MC draws for integral
-        bool probit:         whether to use the probit transformation or not
-        int n_reassignments: number of reassignments. Default is not to reassign.
+        iterable bounds:             boundaries of the rectangle over which the distribution is defined. It should be in the format [[xmin, xmax],[ymin, ymax],...]
+        iterable prior_pars:         IW parameters
+        double alpha0:               initial guess for concentration parameter
+        double MC_draws:             number of MC draws for integral
+        bool probit:                 whether to use the probit transformation or not
+        int n_reassignments:         number of reassignments. Default is not to reassign.
+        callable selection_function: selection function approximant
     
     Returns:
         HDPGMM: instance of HDPGMM class
     """
     def __init__(self, bounds,
-                       alpha0          = 1.,
-                       prior_pars      = None,
-                       MC_draws        = None,
-                       probit          = True,
-                       n_reassignments = 0.,
+                       alpha0             = 1.,
+                       prior_pars         = None,
+                       MC_draws           = None,
+                       probit             = True,
+                       n_reassignments    = 0.,
+                       selection_function = None
                        ):
         bounds   = np.atleast_2d(bounds)
         self.dim = len(bounds)
@@ -1167,6 +1170,7 @@ class HDPGMM(DPGMM):
             self.MC_draws = int((self.dim+1)*1e3)
         else:
             self.MC_draws = int(MC_draws)
+        self.selfunc = selection_function
         self.evaluated_logL = {}
         # MC samples
         self._draw_MC_samples()
@@ -1194,6 +1198,10 @@ class HDPGMM(DPGMM):
             rhos = invwishart(df = self.dim+2, scale = np.identity(self.dim)).rvs(size = self.MC_draws)
             rhos = np.array([r/outer_jit(np.sqrt(diag_jit(r)), np.sqrt(diag_jit(r))) for r in rhos])
             self.sigma_MC = np.array([r*outer_jit(s,s) for r, s in zip(rhos, np.sqrt(self.sigma_MC))])
+        if self.selfunc is not None:
+            self.log_alpha_factor = np.array([np.log(np.mean(self.selfunc(mn(m, s).rvs(1000)))) for m, s in zip(self.mu_MC, self.sigma_MC)])
+        else:
[…]
