+++ figaro/cosmology.py
@@ -62,3 +62,4 @@ class CosmologicalParameters:
             return newton(objective,1.0,args=(self, DL))
 
 Planck18 = CosmologicalParameters(0.674, 0.315, 0.685, -1, 0, 0)
+Planck15 = CosmologicalParameters(0.679, 0.3065, 0.6935, -1, 0, 0)
