### 2024-03-29 bafcccb
added functions to load injections and selection function
