+++ figaro/load.py
@@ -3,6 +3,7 @@ import h5py
 import warnings
 import json
 import dill
+import importlib
 from figaro.exceptions import FIGAROException
 from figaro.mixture import mixture
 from figaro.cosmology import CosmologicalParameters
@@ -11,6 +12,10 @@ from tqdm import tqdm
 
 supported_extensions = ['h5', 'hdf5', 'txt', 'dat', 'csv']
 supported_waveforms  = ['combined', 'imr', 'seob']
+injected_pars        = ['m1', 'm2', 'z', 's1x', 's1y', 's1z', 's2x', 's2y', 's2z', 'ra', 'dec']
+loadable_inj_pars    = injected_pars + ['q', 'chi_eff', 'chi_p', 's1', 's2']
+mass_parameters      = ['m1', 'm2', 'm1_detect', 'm2_detect', 'mc', 'mt', 'q']
+spin_parameters      = ['s1x', 's1y', 's1z', 's2x', 's2y', 's2z', 's1', 's2', 'chi_eff', 'chi_p']
 
 GW_par = {'m1'                 : 'mass_1_source',
           'm2'                 : 'mass_2_source',
@@ -47,6 +52,21 @@ GW_par = {'m1'                 : 'mass_1_source',
           'log_prior'          : 'log_prior',
           'log_likelihood'     : 'log_likelihood',
           }
+
+# LVK 03 injections have a slightly different nomenclature (no underscore)
+# See https://zenodo.org/records/7890437
+inj_par = GW_par
+inj_par['m1']                  = 'mass1_source'
+inj_par['m2']                  = 'mass2_source'
+inj_par['m1_detect']           = 'mass1'
+inj_par['m2_detect']           = 'mass2'
+inj_par['spin1x']              = 'spin1x'
+inj_par['spin1y']              = 'spin1y'
+inj_par['spin1z']              = 'spin1z'
+inj_par['spin2x']              = 'spin2x'
+inj_par['spin2y']              = 'spin2y'
+inj_par['spin2z']              = 'spin2z'
+inj_par['luminosity_distance'] = 'distance'
 
 supported_pars = [p for p in GW_par.keys() if not p in ['snr', 'far']]
 
@@ -89,7 +109,7 @@ def load_single_event(event, seed = False, par = None, n_samples = -1, h = 0.674
         par = ['ra', 'dec', 'luminosity_distance']
     if not ext in supported_extensions:
         raise TypeError("File {0}.{1} is not supported".format(name, ext))
-    if ext == 'txt' or ext == 'dat' or ext == 'csv':
+    if not ext in ['h5', 'hdf5']:
         if par is not None:
             warnings.warn("Par names (or volume keyword) are ignored for .txt/.dat/.csv files")
         if n_samples > -1:
@@ -159,10 +179,9 @@ def load_data(path, seed = False, par = None, n_samples = -1, h = 0.674, om = 0.
         names.append(name)
         if not ext in supported_extensions:
             raise TypeError("File {0}.{1} is not supported".format(name, ext))
-        
-        if ext == 'txt' or ext == 'dat' or ext == 'csv':
+        if not ext in ['h5', 'hdf5']:
             if par is not None:
-                warnings.warn("Par names (or volume keyword) are ignored for .txt/.dat files")
+                warnings.warn("Par names (or volume keyword) are ignored for .txt/.dat/.csv files")
             if n_samples > -1:
                 samples = np.atleast_1d(np.loadtxt(event))
                 s = int(min([n_samples, len(samples)]))
@@ -225,7 +244,7 @@ def _unpack_gw_posterior(event, par, cosmology, rdstate, n_samples = -1, wavefor
[…]
