### 2024-03-29 e16cf06
fixed bug with far_threshold
