### 2024-03-30 019c581
working samples
