### 2024-03-30 98bdafc
more robust numerically - only probit with samples still puzzling
