### 2024-04-05 146042b
fixed bug in averaging samples
