### 2024-04-05 84fe8f7
updated with new selection effects handling
