+++ figaro/_pipelines/par_hierarchical_inference.py
@@ -11,7 +11,7 @@ from figaro.mixture import DPGMM, HDPGMM
 from figaro.transform import transform_to_probit
 from figaro.utils import save_options, load_options, get_priors
 from figaro.plot import plot_median_cr, plot_multidim
-from figaro.load import load_data, load_single_event, save_density, load_density, supported_pars
+from figaro.load import load_data, load_single_event, load_selection_function, save_density, load_density, supported_pars
 
 import ray
 from ray.util import ActorPool
@@ -31,21 +31,25 @@ class worker:
                        save_se     = True,
                        MC_draws    = None,
                        probit      = True,
+                       selfunc     = None,
+                       inj_pdf     = None,
                        ):
         self.dim                  = bounds.shape[0]
         self.bounds               = bounds
         self.mixture              = DPGMM(self.bounds, probit = probit)
         self.hierarchical_mixture = HDPGMM(self.bounds,
-                                           MC_draws   = MC_draws,
-                                           probit     = probit,
-                                           prior_pars = get_priors(self.bounds,
-                                                                   samples      = events,
-                                                                   std          = hier_sigma,
-                                                                   scale        = scale,
-                                                                   probit       = probit,
-                                                                   hierarchical = True,
-                                                                   )
-                                            )
+                                           MC_draws           = MC_draws,
+                                           probit             = probit,
+                                           selection_function = selfunc,
+                                           injection_pdf      = inj_pdf,
+                                           prior_pars         = get_priors(self.bounds,
+                                                                           samples      = events,
+                                                                           std          = hier_sigma,
+                                                                           scale        = scale,
+                                                                           probit       = probit,
+                                                                           hierarchical = True,
+                                                                           ),
+                                           )
         self.out_folder_plots = out_folder_plots
         self.out_folder_draws = out_folder_draws
         self.se_sigma         = se_sigma
@@ -133,8 +137,7 @@ def main():
     parser.add_option("--fraction", dest = "fraction", type = "float", help = "Fraction of samples standard deviation for sigma prior. Overrided by sigma_prior.", default = None)
     parser.add_option("--n_parallel", dest = "n_parallel", type = "int", help = "Number of parallel threads", default = 2)
     parser.add_option("--mc_draws", dest = "mc_draws", type = "int", help = "Number of draws for assignment MC integral", default = None)
-    parser.add_option("--snr_threshold", dest = "snr_threshold", type = "float", help = "SNR threshold for simulated GW datasets", default = None)
-    parser.add_option("--far_threshold", dest = "far_threshold", type = "float", help = "FAR threshold for simulated GW datasets", default = None)
+    parser.add_option("--far_threshold", dest = "far_threshold", type = "float", help = "FAR threshold for LVK sensitivity estimate injections", default = 1.)
     parser.add_option("--no_probit", dest = "probit", action = 'store_false', help = "Disable probit transformation", default = True)
     parser.add_option("--config", dest = "config", type = "string", help = "Config file. Warning: command line options override config options", default = None)
 
@@ -166,6 +169,10 @@ def main():
     # Read hierarchical name
     if options.hier_name is None:
         options.hier_name = options.output.parts[-1]
+    if options.selfunc_file is None:
+        options.hier_name = 'observed_'+options.hier_name
+    else:
+        options.hier_name = 'intrinsic_'+options.hier_name
 
     if options.config is None:
[…]
