### 2024-04-05 618cb8f
deep copy of gw pars dict
