### 2024-04-08 #6589 @editorialbot
[REVIEW]: FIGARO: hierarchical non-parametric inference for population studies
**Submitting author:** @sterinaldi (<a href="http://orcid.org/0000-0001-5799-4155">Stefano Rinaldi</a>)
**Repository:** https://github.com/sterinaldi/FIGARO
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.6.7
