+++ figaro/plot.py
@@ -198,10 +198,11 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
             p_x = injected
         if injected_label is not None:
             injected_label = '$'+injected_label+'$'
-        ax.plot(x, p_x, lw = 0.5, color = 'red', label = injected_label)
         if selfunc is not None:
             filtered_p_x = p_x*f_x
-            ax.plot(x, filtered_p_x/np.sum(filtered_p_x*dx), lw = 0.5, color = 'k', label = '$\mathrm{Selection\ effects}$')
+            ax.plot(x, filtered_p_x/np.sum(filtered_p_x*dx), lw = 0.5, color = 'red', label = '$'+injected_label+'\ \\mathrm{(observed)}$')
+        else:
+            ax.plot(x, p_x, lw = 0.5, color = 'red', label = injected_label)
         
     # Median
     if true_value is not None:
@@ -248,13 +249,13 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
             plot_folder = out_folder
             log_folder  = out_folder
             txt_folder  = out_folder
-        fig.savefig(Path(log_folder, 'log_{0}.pdf'.format(name)), bbox_inches = 'tight')
+        fig.savefig(Path(log_folder, 'log_observed_{0}.pdf'.format(name)), bbox_inches = 'tight')
         ax.set_yscale('linear')
         ax.autoscale(True)
         if samples is not None:
             ax.set_xlim(xlim)
-        fig.savefig(Path(plot_folder, '{0}.pdf'.format(name)), bbox_inches = 'tight')
-        np.savetxt(Path(txt_folder, 'prob_{0}.txt'.format(name)), np.array([x, p[50], p[5], p[16], p[84], p[95]]).T, header = 'x 50 5 16 84 95')
+        fig.savefig(Path(plot_folder, 'observed_{0}.pdf'.format(name)), bbox_inches = 'tight')
+        np.savetxt(Path(txt_folder, 'prob_observed_{0}.txt'.format(name)), np.array([x, p[50], p[5], p[16], p[84], p[95]]).T, header = 'x 50 5 16 84 95')
     if show:
         ax.set_yscale('linear')
         ax.autoscale(True)
@@ -292,11 +293,11 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
         ax.set_ylim(bottom = 1e-5, top = np.max(p[95])*1.1)
         ax.legend(loc = 0)
         if save:
-            fig.savefig(Path(log_folder, 'log_true_{0}.pdf'.format(name)), bbox_inches = 'tight')
+            fig.savefig(Path(log_folder, 'log_intrinsic_{0}.pdf'.format(name)), bbox_inches = 'tight')
             ax.set_yscale('linear')
             ax.autoscale(True)
-            fig.savefig(Path(plot_folder, 'true_{0}.pdf'.format(name)), bbox_inches = 'tight')
-            np.savetxt(Path(txt_folder, 'prob_true_{0}.txt'.format(name)), np.array([x, p[50], p[5], p[16], p[84], p[95]]).T, header = 'x 50 5 16 84 95')
+            fig.savefig(Path(plot_folder, 'intrinsic_{0}.pdf'.format(name)), bbox_inches = 'tight')
+            np.savetxt(Path(txt_folder, 'prob_intrinsic_{0}.txt'.format(name)), np.array([x, p[50], p[5], p[16], p[84], p[95]]).T, header = 'x 50 5 16 84 95')
         if show:
             ax.set_yscale('linear')
             ax.autoscale(True)
