### 2024-04-10 60f3395
added keywords for colors and colormap
