+++ figaro/plot.py
@@ -99,7 +99,7 @@ class PPPlot(axes.Axes):
         
 projection_registry.register(PPPlot)
 
-def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bounds = None, out_folder = '.', name = 'density', n_pts = 1000, label = None, unit = None, hierarchical = False, show = False, save = True, subfolder = False, true_value = None, true_value_label = '\mathrm{True\ value}', injected_label = '\mathrm{Simulated}', median_label = None, fig = None):
+def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bounds = None, out_folder = '.', name = 'density', n_pts = 1000, label = None, unit = None, hierarchical = False, show = False, save = True, subfolder = False, true_value = None, true_value_label = '\mathrm{True\ value}', injected_label = '\mathrm{Simulated}', median_label = None, fig = None, colors = ['steelblue', 'darkturquoise', 'mediumturquoise']):
     """
     Plot the recovered 1D distribution along with the injected distribution and samples from the true distribution (both if available).
     
@@ -123,6 +123,8 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
         str injected_label:              label to assign to the injected distribution
         str median_label:                label to assign to the reconstruction
         matplotlib.figure.Figure fig:    figure to use for plotting. Must have (dim,dim) axes.
+        list-of-str colors:              list of colors for median, 68% and 90% credible regions
+        
     
     Returns:
         matplotlib.figure.Figure: figure with the plot
@@ -132,6 +134,8 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
             median_label = '\mathrm{(H)DPGMM}'
         else:
             median_label = '\mathrm{DPGMM}'
+    
+    color_med, color_68, color_90 = colors
     
     all_bounds = np.atleast_2d([d.bounds[0] for d in draws])
     x_min = np.max(all_bounds[:,0])
@@ -182,8 +186,8 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
         ax.set_yscale('log')
     
     # CR
-    ax.fill_between(x, p[95], p[5], color = 'mediumturquoise', alpha = 0.25)
-    ax.fill_between(x, p[84], p[16], color = 'darkturquoise', alpha = 0.25)
+    ax.fill_between(x, p[95], p[5], color = color_90, alpha = 0.25)
+    ax.fill_between(x, p[84], p[16], color = color_68, alpha = 0.25)
     # Selection function (if available)
     if selfunc is not None:
         if callable(selfunc):
@@ -208,8 +212,8 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
     if true_value is not None:
         if true_value_label is not None:
             true_value_label = '$'+true_value_label+'$'
-        ax.axvline(true_value, ls = '--', color = 'r', dashes = (5,5), lw = 0.5, label = true_value_label)
-    ax.plot(x, p[50], lw = 0.7, color = 'steelblue', label = '${0}$'.format(median_label))
+        ax.axvline(true_value, ls = '--', color = 'firebrick', dashes = (5,5), lw = 0.5, label = true_value_label)
+    ax.plot(x, p[50], lw = 0.7, color = color_med, label = '${0}$'.format(median_label))
     if label is None:
         label = 'x'
     if unit is None or unit == '':
@@ -305,7 +309,7 @@ def plot_median_cr(draws, injected = None, samples = None, selfunc = None, bound
         plt.close()
     return fig
 
[…]
