### 2024-04-10 7749ba3
naming convenction of 1d plots
