### 2024-04-12 3e1997d
checks if self.selfunc is not None before evaluating logpdf
