+++ pyproject.toml
@@ -13,7 +13,7 @@ packages = ['figaro', 'figaro._pipelines']
 [project]
 name = 'figaro'
 description = 'FIGARO: Fast Inference for GW Astronomy, Research & Observations'
-version = '1.5.5'
+version = '1.6.0'
 requires-python = '< 3.12'
 readme = {file = "README.md", content-type = "text/markdown"}
 license = {file = "LICENSE"}
