### 2024-04-15 2ab0cb0
fixed bug with n_injections
