+++ figaro/_pipelines/hierarchical_inference.py
@@ -112,8 +112,9 @@ def main():
         spec.loader.exec_module(inj_module)
         inj_density = inj_module.density
     # If provided, load selecton function
-    selfunc = None
-    inj_pdf = None
+    selfunc     = None
+    inj_pdf     = None
+    n_total_inj = None
     if options.selfunc_file is not None:
         selfunc, inj_pdf, n_total_inj = load_selection_function(options.selfunc_file, par = options.par, far_threshold = options.far_threshold)
         if not callable(selfunc) and not options.probit:
