### 2024-04-15 4f6b9eb
log scale for n_true
