### 2024-04-15 6560d2a
check if cov matrix is zero
