+++ figaro/_likelihood.py
@@ -71,10 +71,15 @@ def log_norm_int(x, mu, cov_1, inv_cov_1, cov_2):
     Returns:
         double: MultivariateNormal(m,s).logpdf(x)
     """
-    inv_cov_2  = inv_jit(cov_2)
-    inv_cov    = inv_cov_1@inv_jit(inv_cov_1+inv_cov_2)@inv_cov_2
+    if not np.all(cov_2 == 0.):
+        inv_cov_2  = inv_jit(cov_2)
+        inv_cov    = inv_cov_1@inv_jit(inv_cov_1+inv_cov_2)@inv_cov_2
+        logdet     = (logdet_jit(cov_1) + logdet_jit(cov_2) + logdet_jit(inv_cov_1+inv_cov_2))
+    else:
+        inv_cov = inv_cov_1
+        logdet  = logdet_jit(cov_1)
     exponent   = -0.5*scalar_product(x-mu, inv_cov, len(mu))
-    lognorm    = 0.5*len(mu)*LOG2PI+0.5*(logdet_jit(cov_1) + logdet_jit(cov_2) + logdet_jit(inv_cov_1+inv_cov_2))
+    lognorm    = 0.5*len(mu)*LOG2PI+0.5*logdet
     return -lognorm+exponent
 
 #------------#
