### 2024-04-16 ae1efe1
storing alpha factor
