+++ figaro/mixture.py
@@ -275,7 +275,7 @@ class _component_h:
         if dim == 1:
             self.mu = np.atleast_2d(self.mu).T
             self.sigma = np.atleast_2d(self.sigma).T
-        self.N_true = np.exp(np.log(self.N) - log_alpha_factor[idx])
+        self.log_N_true = np.log(self.N) - log_alpha_factor[idx]
             
 class density:
     """
@@ -681,22 +681,23 @@ class mixture(density):
     Methods inherited from density class
     
     Arguments:
-        iterable means:    component means
-        iterable covs:     component covariances
-        np.ndarray w:      component weights
-        np.ndarray bounds: bounds of probit transformation
-        int dim:           number of dimensions
-        int n_cl:          number of clusters in the mixture
-        int n_pts:         number of points used to infer the mixture
-        double alpha:      concentration parameter
-        bool probit:       whether to use the probit transformation or not
-        np.ndarray log_w:  component log weights
-        bool make_comp:    make component objects
+        iterable means:      component means
+        iterable covs:       component covariances
+        np.ndarray w:        component weights
+        np.ndarray bounds:   bounds of probit transformation
+        int dim:             number of dimensions
+        int n_cl:            number of clusters in the mixture
+        int n_pts:           number of points used to infer the mixture
+        double alpha:        concentration parameter
+        bool probit:         whether to use the probit transformation or not
+        np.ndarray log_w:    component log weights
+        bool make_comp:      make component objects
+        double alpha_factor: evaluated \int pdet(theta)p(theta|lambda) conditioned on the mixture parameters
     
     Returns:
         mixture: instance of mixture class
     """
-    def __init__(self, means, covs, w, bounds, dim, n_cl, n_pts, alpha = 1., probit = True, log_w = None, make_comp = True):
+    def __init__(self, means, covs, w, bounds, dim, n_cl, n_pts, alpha = 1., probit = True, log_w = None, make_comp = True, alpha_factor = 1.):
         self.means = means
         self.covs  = covs
         if make_comp:
@@ -708,13 +709,14 @@ class mixture(density):
             self.log_w  = np.log(w)
         else:
             self.log_w  = log_w
-            self.w      = np.exp(log_w)
-        self.bounds = bounds
-        self.dim    = int(dim)
-        self.n_cl   = int(n_cl)
-        self.n_pts  = int(n_pts)
-        self.probit = probit
-        self.alpha  = alpha
+            self.w        = np.exp(log_w)
+        self.bounds       = bounds
+        self.dim          = int(dim)
+        self.n_cl         = int(n_cl)
+        self.n_pts        = int(n_pts)
+        self.probit       = probit
+        self.alpha        = alpha
+        self.alpha_factor = alpha_factor
     
     def marginalise(self, axis = -1):
         """
@@ -1182,9 +1184,10 @@ class HDPGMM(DPGMM):
                 self.total_inj   = int(total_injections)
             except TypeError:
                 raise FIGAROException("Please provide injection pdf")
+            self.selfunc_probit   = self.selfunc
             self.log_jacobian_inj = np.zeros(len(self.selfunc))
             if self.probit:
-                self.selfunc = transform_to_probit(self.selfunc, self.bounds)
+                self.selfunc_probit   = transform_to_probit(self.selfunc, self.bounds)
[…]
