### 2024-04-16 747c798
added alpha_factor
