### 2024-04-16 28744b6
added rate in plot_median_cr
