+++ figaro/rate.py
@@ -1,0 +1,262 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from pathlib import Path
+from collections import Counter
+from scipy.stats import poisson, norm
+from figaro.cosmology import dVdz_approx_planck18, Planck18 as omega
+
+def sample_rate(draws, n_obs, selfunc, T, volume = None, size = None, each = False, n_draws = 1e4, dvdz = None, z_index = -1, normalise_alpha = False):
+    """
+    Compute the integrated rate given a set of draws. If volume keyword is not used, returns the expected number of events before filtering.
+    
+    Arguments:
+        iterable draws:       container for mixture instances
+        int n_obs:            number of observed events
+        callable selfunc:     selection function approximant
+        double T:             duration of the observations
+        double volume:        surveyed volume. If None, it is estimated.
+        int size:             number of samples
+        bool each:            sample one point from each draw
+        int n_draws:          number of MC draws
+        callable dvdz         function to compute the comoving volume element (in Gpc^3)
+        int z_index           redshift parameter index (default: last dimension)
+        bool normalise_alpha: re-normalise alpha (required if the selection function includes dV/dz*(1+z)^-1)
+    
+    Returns:
+        np.ndarray: integrated rate samples
+    """
+    if size is None or each:
+        size = len(draws)
+    if volume is None:
+        volume = sample_VT(draws, selfunc, T, size = size, each = each, dvdz = dvdz, z_index = z_index)
+    if normalise_alpha:
+        normalise_alpha_factor(draws, dvdz = dvdz, z_index = z_index, n_draws = n_draws)
+    N_exp   = [poisson(int(n_obs/d.alpha_factor)) for d in draws]
+    if each:
+        return np.array([N.rvs()/v for N, v in zip(N_exp, volume)])
+    idx     = np.random.choice(np.arange(len(N_exp)), size = int(size))
+    ctr     = Counter(idx)
+    samples = np.empty(shape = (1,))
+    for i, n in zip(ctr.keys(), ctr.values()):
+        samples = np.concatenate((samples, N_exp[i].rvs(n)))
+    return np.divide(samples[1:], volume)
+
+def normalise_alpha_factor(draws, dvdz = None, z_index = -1, n_draws = 1e4):
+    """
+    Normalise the alpha factor for each DPGMM realisation. This is required if during the reconstruction the selection function included the dV/dz*(1+z)^-1 term.
+    
+    Arguments:
+        iterable draws: container for mixture instances
+        callable dvdz:  function to compute the comoving volume element (in Gpc^3)
+        int z_index:    redshift parameter index (default: last dimension)
+        int n_draws:    number of MC draws
+    """
+    if dvdz is None:
+        dvdz = dVdz_approx_planck18
+    for d in draws:
+        ss = d.rvs(int(n_draws))
+        d.alpha_factor /= np.mean(dvdz(ss[:,z_index])/(1+ss[:,z_index]))
+
+def sample_VT(draws, selfunc, T, size = None, n_draws = 1e4, each = False, dvdz = None, z_index = -1):
+    """
+    Draw samples from the VT distribution given the astrophysical distribution of quantities, the selection function and the observing time.
+    See Eqs. (21) of Kapadia et al (2020) (https://arxiv.org/pdf/1903.06881.pdf).
+    The calibration uncertainty is estimated as in Sec. 5 - Eq. (17) - of Abbott et al (2016) (https://arxiv.org/pdf/1606.03939.pdf)
+    
+    Arguments:
+        iterable draws:   container for mixture instances
+        callable selfunc: selection function approximant
[…]
