+++ figaro/cosmology.py
@@ -2,6 +2,7 @@ import numpy as np
 import warnings
 warnings.filterwarnings("ignore", "Wswiglal-redir-stdio") # Silence LAL warnings with ipython
 from scipy.optimize import newton
+from scipy.interpolate import interp1d
 from lal._lal import LuminosityDistance, UniformComovingVolumeDensity, ComovingVolumeElement, ComovingVolume, CreateCosmologicalParameters
 
 class CosmologicalParameters:
@@ -63,3 +64,8 @@ class CosmologicalParameters:
 
 Planck18 = CosmologicalParameters(0.674, 0.315, 0.685, -1, 0, 0)
 Planck15 = CosmologicalParameters(0.679, 0.3065, 0.6935, -1, 0, 0)
+
+# Interpolants up to z = 2
+z = np.linspace(0,2,1000)
+dVdz_approx_planck18 = interp1d(z, Planck18.ComovingVolumeElement(z)/1e9) # In Gpc
+dVdz_approx_planck15 = interp1d(z, Planck15.ComovingVolumeElement(z)/1e9) # In Gpc
