### 2024-04-17 2bc797e
load_selection_function returns duration and load_data/load_single_event requires name of cosmological parameters set
