+++ figaro/load.py
@@ -7,7 +7,7 @@ import copy
 import importlib
 from figaro.exceptions import FIGAROException
 from figaro.mixture import mixture
-from figaro.cosmology import CosmologicalParameters
+from figaro.cosmology import Planck18, Planck15
 from pathlib import Path
 from tqdm import tqdm
 
@@ -77,22 +77,20 @@ def available_gw_pars():
     """
     print(supported_pars)
 
-def load_single_event(event, seed = False, par = None, n_samples = -1, h = 0.674, om = 0.315, ol = 0.685, volume = False, waveform = 'combined', snr_threshold = None, far_threshold = None):
+def load_single_event(event, seed = False, par = None, n_samples = -1, cosmology = 'Planck18', volume = False, waveform = 'combined', snr_threshold = None, far_threshold = None):
     '''
     Loads the data from .txt/.dat files (for simulations) or .h5/.hdf5 files (posteriors from GWTC) for a single event.
     Default cosmological parameters from Planck Collaboration (2021) in a flat Universe (https://www.aanda.org/articles/aa/pdf/2020/09/aa33910-18.pdf)
     Not all GW parameters are implemented: run figaro.load.available_gw_pars() for a list of the available parameters.
     
     Arguments:
-        str or Path file: file with samples
-        bool seed:        fixes the seed to a default value (1) for reproducibility
-        list-of-str par:  list with parameter(s) to extract from GW posteriors (m1, m2, mc, z, chi_effective)
-        int n_samples:    number of samples for (random) downsampling. Default -1: all samples
-        double h:         Hubble constant H0/100 [km/(s*Mpc)]
-        double om:        matter density parameter
-        double ol:        cosmological constant density parameter
-        bool volume:      if True, loads RA, dec and Luminosity distance (for skymaps)
-        str waveform:     waveform family to be used ('combined', 'seob', 'imr')
+        str or Path file:     file with samples
+        bool seed:            fixes the seed to a default value (1) for reproducibility
+        list-of-str par:      list with parameter(s) to extract from GW posteriors (m1, m2, mc, z, chi_effective)
+        int n_samples:        number of samples for (random) downsampling. Default -1: all samples
+        str cosmology:        set of cosmological parameters (Planck18 or Planck15)
+        bool volume:          if True, loads RA, dec and Luminosity distance (for skymaps)
+        str waveform:         waveform family to be used ('combined', 'seob', 'imr')
         double snr_threhsold: SNR threshold for event filtering. For injection analysis only.
         double far_threshold: FAR threshold for event filtering. For injection analysis only.
         
@@ -129,7 +127,7 @@ def load_single_event(event, seed = False, par = None, n_samples = -1, h = 0.674
             raise FIGAROException("The following parameters are not implemented: "+", ".join(wrong_pars)+". Run figaro.load.available_gw_pars() for a list of available parameters.")
         # If everything is ok, load the samples
         else:
-            out = _unpack_gw_posterior(event, par = par, n_samples = n_samples, cosmology = (h, om, ol), rdstate = rdstate, waveform = waveform, snr_threshold = snr_threshold, far_threshold = far_threshold)
+            out = _unpack_gw_posterior(event, par = par, n_samples = n_samples, cosmology = cosmology, rdstate = rdstate, waveform = waveform, snr_threshold = snr_threshold, far_threshold = far_threshold)
     
     if out is None:
         return out, name
@@ -138,7 +136,7 @@ def load_single_event(event, seed = False, par = None, n_samples = -1, h = 0.674
         out = np.atleast_2d(out).T
     return out, name
 
[…]
