### 2024-04-17 fbcda7b
returns best estimate for rate if each = True
