### 2024-04-17 0b6205e
updated load_selection_function
