+++ figaro/_pipelines/par_hierarchical_inference.py
@@ -12,6 +12,9 @@ from figaro.transform import transform_to_probit
 from figaro.utils import save_options, load_options, get_priors
 from figaro.plot import plot_median_cr, plot_multidim
 from figaro.load import load_data, load_single_event, load_selection_function, save_density, load_density, supported_pars
+from figaro.rate import sample_rate, normalise_alpha_factor, plot_integrated_rate, plot_differential_rate
+from figaro.cosmology import _decorator_dVdz, dVdz_approx_planck18, dVdz_approx_planck15
+from figaro.marginal import marginalise
 
 import ray
 from ray.util import ActorPool
@@ -132,7 +135,7 @@ def main():
     parser.add_option("--se_draws", type = "int", dest = "se_draws", help = "Number of draws for single-event distribution. Default: same as hierarchical distribution", default = None)
     parser.add_option("--n_samples_dsp", type = "int", dest = "n_samples_dsp", help = "Number of samples to analyse (downsampling). Default: all", default = -1)
     parser.add_option("--exclude_points", dest = "exclude_points", action = 'store_true', help = "Exclude points outside bounds from analysis", default = False)
-    parser.add_option("--cosmology", type = "string", dest = "cosmology", help = "Cosmological parameters (h, om, ol). Default values from Planck (2021)", default = '0.674,0.315,0.685')
+    parser.add_option("--cosmology", type = "choice", dest = "cosmology", help = "Set of cosmological parameters. Default values from Planck (2021)", choices = ['Planck18', 'Planck15'] default = 'Planck18')
     parser.add_option("-e", "--events", dest = "run_events", action = 'store_false', help = "Skip single-event analysis", default = True)
     parser.add_option("--se_sigma_prior", dest = "se_sigma_prior", type = "string", help = "Expected standard deviation (prior) for single-event inference - single value or n-dim values. If None, it is estimated from samples", default = None)
     parser.add_option("--sigma_prior", dest = "sigma_prior", type = "string", help = "Expected standard deviation (prior) for hierarchical inference - single value or n-dim values. If None, it is estimated from samples", default = None)
@@ -142,6 +145,8 @@ def main():
     parser.add_option("--far_threshold", dest = "far_threshold", type = "float", help = "FAR threshold for LVK sensitivity estimate injections", default = 1.)
     parser.add_option("--no_probit", dest = "probit", action = 'store_false', help = "Disable probit transformation", default = True)
     parser.add_option("--config", dest = "config", type = "string", help = "Config file. Warning: command line options override config options", default = None)
+    parser.add_option("--rate", dest = "rate", action = 'store_true', help = "Compute rate", default = False)
+    parser.add_option("--include_dvdz", dest = "include_dvdz", action = 'store_true', help = "Include dV/dz*(1+z)^{-1} term in selection effects.", default = False)
 
     (options, args) = parser.parse_args()
 
@@ -168,6 +173,10 @@ def main():
     output_draws = Path(options.output, 'draws')
     if not output_draws.exists():
         output_draws.mkdir()
+    if options.rate:
+        output_rate = Path(options.output, 'rate')
+        if not output_rate.exists():
+            output_rate.mkdir()
     # Read hierarchical name
     if options.hier_name is None:
         options.hier_name = options.output.parts[-1]
@@ -184,9 +193,6 @@ def main():
         options.bounds = np.array(np.atleast_2d(eval(options.bounds)), dtype = np.float64)
     elif options.bounds is None and not options.postprocess:
         raise Exception("Please provide bounds for the inference (use -b '[[xmin,xmax],[ymin,ymax],...]')")
-
-    # Read cosmology
[…]
