### 2024-04-18 83f1f37
added regularisation for dvdz
