### 2024-04-18 d3161ae
import duration from selfunc_module
