### 2024-04-18 14a748d
changed numerical stability value
