### 2024-04-18 c8bc43c
allow singular in alpha computing
