### 2024-04-22 41d4db1
fig in plot_1d_dist and maquillage
