### 2024-04-23 #6589 @dgerosa
Hi @sterinaldi, here are my comments following the checklist above:

- The statement of need is not reported in the documentation.
- The package has several dependencies, as specified in `pyproject.toml`. I've installed the code in a fresh environment and confirm this works, but those dependencies should be listed in the documentation.
- Please specify if you're open to community contributions and how (e.g. open an issue / pull request on github).
- I could not find automated tests. Please provide some simple unit testing infrastructure, perhaps fitting simple distribution you know the analytic answer of.
- The documentation could be improved with plots instead of just instructions. I think this will make the package more attractive.
- Following up on this, I would actually embed the notebook `introductive_guide.ipynb` into a readthedocs page (this task can be automated with Sphinx) so that a casual user can see what the code is like without cloning the full repository, running the code, etc.
- The paper does not describe how this software compares with other software that tackles the same problem. There is a large industry of population fits in gravitational-wave astronomy, which should be mentioned/discussed/cited.
- The paper could be expanded with a tiny bit more details on the maths behind Dirichlet Processes.
- To ensure the documentation stays up to date with the code, I suggest setting up a github action.
