### 2024-04-23 #6589 @dgerosa
I also have a comment/question for @dfm, which is about the scope of JOSS. The authors already published a paper in the scientific literature that presents their FIGARO code:

Monthly Notices of the Royal Astronomical Society: Letters, Volume 517, Issue 1, November 2022, Pages L5–L10, https://doi.org/10.1093/mnrasl/slac101

This JOSS paper also presents the code and their other paper provides more details. Should there be a JOSS paper for this code? Should this submission be considered a duplicate?
