+++ docs/source/figaro.rate.rst
@@ -1,0 +1,7 @@
+figaro.load module
+==================
+
+.. automodule:: figaro.rate
+   :members:
+   :undoc-members:
+   :show-inheritance:
