### 2024-04-26 5d3fa18
plt.close only if save or show
