### 2024-05-05 #6589 @dfm
@sterinaldi — Thanks for your response above and I'm so sorry about my delayed reply. This does sound to me like this submission is sufficiently different from the MNRAS publication to be consistent with the JOSS requirements. With this in mind, I think we should continue with the review. Thanks again to @dgerosa for checking in on this point!
