+++ docs/source/api_guide.rst
@@ -1,0 +1,4 @@
+.. toctree::
+   :maxdepth: 1
+    
+    ../../introductive_guide.ipynb
