+++ docs/source/api_guide.rst
@@ -1,3 +1,6 @@
+API guide
+=========
+
 .. toctree::
    :maxdepth: 1
     
