+++ docs/source/index.rst
@@ -29,7 +29,7 @@ FIGARO has the following dependencies:
 * lalsuite
 * ray
 
-../../introductive_guide.ipynb
+introductive_guide.ipynb
 
 .. toctree::
    :maxdepth: 1
