+++ docs/source/api_guide.rst
@@ -4,4 +4,4 @@ API guide
 .. toctree::
    :maxdepth: 1
     
-    ../../introductive_guide.ipynb
+    introductive_guide.ipynb
