+++ docs/requirements.txt
@@ -2,4 +2,4 @@ myst_parser
 sphinx-rtd-theme
 figaro
 nbsphinx
-ipython
+ipykernel
