+++ docs/source/figaro.rst
@@ -14,5 +14,6 @@ Documentation
    figaro.mixture
    figaro.montecarlo
    figaro.plot
+   figaro.rate
    figaro.transform
    figaro.utils
