+++ docs/source/index.rst
@@ -29,8 +29,6 @@ FIGARO has the following dependencies:
 * lalsuite
 * ray
 
-introductive_guide.ipynb
-
 .. toctree::
    :maxdepth: 1
    :caption: Contents:
