+++ docs/source/api_guide.rst
@@ -3,4 +3,4 @@ API guide
 
 .. toctree::
    :maxdepth: 1
-   introductive_guide
+   introductive_nb
