### 2024-05-10 7d98fa4
flattened 1-dim rvs
