### 2024-05-10 7757040
added use_mixture notebook
