+++ docs/source/api_guide.rst
@@ -1,7 +1,0 @@
-FIGARO in a python script
-=========
-
-.. toctree::
-   :maxdepth: 2
-   
-   ./introductive_notebook
