### 2024-05-13 dfa06f6
added likelihood sampling
