### 2024-05-14 130c7f4
added mass priors and jacobians
