+++ docs/source/installation.md
@@ -13,3 +13,17 @@ git clone git@github.com:sterinaldi/FIGARO.git
 cd FIGARO
 pip install .
 ```
+## Dependencies
+
+FIGARO has the following dependencies:
+
+* numpy > 1.22
+* scipy
+* matplotlib
+* dill
+* Corner
+* Numba
+* h5py
+* tqdm
+* lalsuite
+* ray
