+++ docs/source/index.rst
@@ -6,7 +6,7 @@
 FIGARO - Fast Inference for GW Astronomy, Research & Observations
 =================================================================
 
-| This page contains the documentation for `FIGARO <https://github.com/sterinaldi/FIGARO>_`, along with a brief description of how to use it in your research project. Please refer to the `GitHub issue tracker <https://github.com/sterinaldi/FIGARO/issues>`_ for bug reports (extremely appreciated), issues and contributions.
+| This page contains the documentation for `FIGARO <https://github.com/sterinaldi/FIGARO>`_, along with a brief description of how to use it in your research project. Please refer to the `GitHub issue tracker <https://github.com/sterinaldi/FIGARO/issues>`_ for bug reports (extremely appreciated), issues and contributions.
 | If you're curious about the other projects I'm working on or if you want to get in touch with me, feel free to visit `my website <https://sterinaldi.github.io>`_!
 
 Statement of need
