+++ paper.bib
@@ -35,18 +35,23 @@ archivePrefix = {arXiv},
 }
 
 @article{rinaldi:2024:m1qz,
-       author = {{Rinaldi}, Stefano and {Del Pozzo}, Walter and {Mapelli}, Michela and {Lorenzo Medina}, Ana and {Dent}, Thomas},
-       title  = "{Evidence for the evolution of black hole mass function with redshift}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
-         year = 2023,
-        month = oct,
-          eid = {arXiv:2310.03074},
-        pages = {arXiv:2310.03074},
+       author = {{Rinaldi}, Stefano and {Del Pozzo}, Walter and {Mapelli}, Michela and {Lorenzo-Medina}, Ana and {Dent}, Thomas},
+        title = "{Evidence of evolution of the black hole mass function with redshift}",
+      journal = {Astronomy \& Astrophysics},
+     keywords = {gravitation, gravitational waves, stars: black holes, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Astrophysics of Galaxies, General Relativity and Quantum Cosmology},
+         year = 2024,
+        month = apr,
+       volume = {684},
+          eid = {A204},
+        pages = {A204},
+          doi = {10.1051/0004-6361/202348161},
 archivePrefix = {arXiv},
        eprint = {2310.03074},
  primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...684A.204R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @ARTICLE{cheung:2023,
        author = {{Cheung}, Damon H.~T. and {Rinaldi}, Stefano and {Toscani}, Martina and {Hannuksela}, Otto A.},
@@ -186,4 +191,23 @@ archivePrefix = {arXiv},
   pages   = {261--272},
   adsurl  = {https://rdcu.be/b08Wh},
   doi     = {10.1038/s41592-019-0686-2},
+}
+
+@ARTICLE{astrodistGWTC3:2023,
+       author = {{Abbott}, R. and others},
+        title = "{Population of Merging Compact Binaries Inferred Using Gravitational Waves through GWTC-3}",
+      journal = {Physical Review X},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
+         year = 2023,
+        month = jan,
+       volume = {13},
+       number = {1},
+          eid = {011048},
+        pages = {011048},
+          doi = {10.1103/PhysRevX.13.011048},
+archivePrefix = {arXiv},
+       eprint = {2111.03634},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023PhRvX..13a1048A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
