### 2024-05-15 e1279b3
added DPGMM details
