+++ paper.bib
@@ -211,3 +211,104 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023PhRvX..13a1048A},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{tiwari:2021:vamana,
+       author = {{Tiwari}, Vaibhav},
+        title = "{VAMANA: modeling binary black hole population with minimal assumptions}",
+      journal = {Classical and Quantum Gravity},
+     keywords = {gravitational waves, population properties, mass distribution, binary black holes, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2021,
+        month = aug,
+       volume = {38},
+       number = {15},
+          eid = {155007},
+        pages = {155007},
+          doi = {10.1088/1361-6382/ac0b54},
+archivePrefix = {arXiv},
+       eprint = {2006.15047},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021CQGra..38o5007T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{edelman:2023,
+       author = {{Edelman}, Bruce and {Farr}, Ben and {Doctor}, Zoheyr},
+        title = "{Cover Your Basis: Comprehensive Data-driven Characterization of the Binary Black Hole Population}",
+      journal = {\apj},
+     keywords = {Gravitational wave astronomy, Gravitational waves, Black holes, Compact objects, High energy astrophysics, 675, 678, 162, 288, 739, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Astrophysics of Galaxies},
+         year = 2023,
+        month = mar,
+       volume = {946},
+       number = {1},
+          eid = {16},
+        pages = {16},
+          doi = {10.3847/1538-4357/acb5ed},
+archivePrefix = {arXiv},
+       eprint = {2210.12834},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...946...16E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{toubiana:2023,
+       author = {{Toubiana}, A. and {Katz}, Michael L. and {Gair}, Jonathan R.},
+        title = "{Is there an excess of black holes around 20 M{\ensuremath{\odot}}? Optimizing the complexity of population models with the use of reversible jump MCMC.}",
+      journal = {\mnras},
+     keywords = {black hole physics, gravitational waves, methods: statistical, General Relativity and Quantum Cosmology, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2023,
+        month = oct,
+       volume = {524},
+       number = {4},
+        pages = {5844-5853},
+          doi = {10.1093/mnras/stad2215},
+archivePrefix = {arXiv},
+       eprint = {2305.08909},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023MNRAS.524.5844T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{callister:2024,
+       author = {{Callister}, Thomas A. and {Farr}, Will M.},
+        title = "{Parameter-Free Tour of the Binary Black Hole Population}",
+      journal = {Physical Review X},
[…]
