+++ paper.bib
@@ -194,7 +194,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{astrodistGWTC3:2023,
-       author = {{Abbott}, R. and others},
+       author = {{Abbott}, R. and {Abbott}, T.~D. and {Acernese}, F. and {Ackley}, K. and {Adams}, C. and {Adhikari}, N. and {Adhikari}, R.~X. and {Adya}, V.~B. and {Affeldt}, C. and {Agarwal}, D. and {Agathos}, M. and {Agatsuma}, K. and {Aggarwal}, N. and {Aguiar}, O.~D. and {Aiello}, L. and {Ain}, A. and {Ajith}, P. and {Akutsu}, T. and {de Alarc{\'o}n}, P.~F. and {Akcay}, S. and {Albanesi}, S. and {Allocca}, A. and {Altin}, P.~A. and {Amato}, A. and {Anand}, C. and {Anand}, S. and {Ananyeva}, A. and {Anderson}, S.~B. and {Anderson}, W.~G. and {Ando}, M. and {Andrade}, T. and {Andres}, N. and {Andri{\'c}}, T. and {Angelova}, S.~V. and {Ansoldi}, and others},
         title = "{Population of Merging Compact Binaries Inferred Using Gravitational Waves through GWTC-3}",
       journal = {Physical Review X},
      keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
