+++ paper.md
@@ -39,7 +39,7 @@ The astrophysical graveyard is populated by black holes (BHs) and neutron stars 
 Bayesian non-parametric methods are models with a countably infinite number of parameters to model arbitrary probability densities. These parameters do not have any connection with the modelled distribution, making these a convenient and agnostic way of describing some unknown population. These are key tools to reconstruct probability densities without being committal to a specific functional form: the basic idea is to let the data speak for themselves, retrieving the distribution that is the most likely to have generated the observed data.
 In a certain sense, this is the most phenomenological approach possible: unlike the standard parametric approach, where we specify a functional form inspired by what we expect to find in the data, with non-parametric methods all the information comes from the data, thus avoiding the risk of biasing the inference with inaccurate models. Features in the inferred distribution will arise naturally without the need of including them in the model, leaving astrophysicists tasked with explaining them in terms of formation channels and astrophysical processes. 
 
-The GW community is currently exploring this direction [e.g. @tiwari:2021:vamana @edelman:2023 @toubiana:2023 @callister:2024]: `FIGARO` fits in this framework, being a non-parametric inference scheme designed to reconstruct arbitrary probability densities in a hierarchical fashion under the requirement of minimal mathematical assumptions.
+The GW community is currently exploring this direction [e.g. @tiwari:2021:vamana; @edelman:2023; @toubiana:2023; @callister:2024]: `FIGARO` fits in this framework, being a non-parametric inference scheme designed to reconstruct arbitrary probability densities in a hierarchical fashion under the requirement of minimal mathematical assumptions.
 
 # Statement of need
 
