+++ paper.bib
@@ -154,19 +154,25 @@ archivePrefix = {arXiv},
  publisher = {[American Statistical Association, Taylor & Francis, Ltd.]},
  title = {Bayesian Density Estimation and Inference Using Mixtures},
  volume = {90},
- year = {1995}
+ year = {1995},
+ doi = {10.1080/01621459.1995.10476550},
 }
 
 @article{pedregosa:2011,
- title={Scikit-learn: Machine Learning in {P}ython},
- author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
-         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
-         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
-         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
- journal={Journal of Machine Learning Research},
- volume={12},
- pages={2825--2830},
- year={2011}
+       author = {{Pedregosa}, Fabian and {Varoquaux}, Ga{\"e}l and {Gramfort}, Alexandre and {Michel}, Vincent and {Thirion}, Bertrand and {Grisel}, Olivier and {Blondel}, Mathieu and {M{\"u}ller}, Andreas and {Nothman}, Joel and {Louppe}, Gilles and {Prettenhofer}, Peter and {Weiss}, Ron and {Dubourg}, Vincent and {Vanderplas}, Jake and {Passos}, Alexandre and {Cournapeau}, David and {Brucher}, Matthieu and {Perrot}, Matthieu and {Duchesnay}, {\'E}douard},
+        title = "{Scikit-learn: Machine Learning in Python}",
+      journal = {Journal of Machine Learning Research},
+     keywords = {Python, supervised learning, unsupervised learning, model selection, Computer Science - Machine Learning, Computer Science - Mathematical Software},
+         year = 2011,
+        month = oct,
+       volume = {12},
+        pages = {2825-2830},
+          doi = {10.48550/arXiv.1201.0490},
+archivePrefix = {arXiv},
+       eprint = {1201.0490},
+ primaryClass = {cs.LG},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011JMLR...12.2825P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @ARTICLE{virtanen:2020,
@@ -285,19 +291,6 @@ archivePrefix = {arXiv},
  primaryClass = {astro-ph.HE},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024PhRvX..14b1005C},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{escobar:1995,
- ISSN = {01621459},
- URL = {http://www.jstor.org/stable/2291069},
- author = {Michael D. Escobar and Mike West},
- journal = {Journal of the American Statistical Association},
- number = {430},
- pages = {577--588},
- publisher = {[American Statistical Association, Taylor & Francis, Ltd.]},
- title = {Bayesian Density Estimation and Inference Using Mixtures},
- volume = {90},
- year = {1995}
 }
 
 @Inbook{teh:2010,
