### 2024-05-15 #6589 @sterinaldi
Hi @dgerosa, sorry for my late reply. Thanks for your review, below you find my reply and changes:

* I added the statement of need in [the home page of the documentation](https://figaro.readthedocs.io/en/latest/#statement-of-need);
* Dependencies are now listed in the [Installation page of the documentation](https://figaro.readthedocs.io/en/latest/installation.html#dependencies);
* The repository now contains a [CONTRIBUTING](https://github.com/sterinaldi/FIGARO/blob/main/CONTRIBUTING.md) file, as well as specifying how to report bugs and issues in the home page of the documentation;
* I added two tests, one for the DPGMM reconstruction and the other for the (H)DPGMM inference: they can be found in [this file](https://github.com/sterinaldi/FIGARO/blob/main/test/test_figaro.py). `python -m unittest` from the main folder of the repository runs them;
* Two notebooks are now embedded in the documentation ([how to use a FIGARO reconstruction](https://figaro.readthedocs.io/en/latest/use_mixture.html) and [how to include FIGARO in a python script](https://figaro.readthedocs.io/en/latest/python_script.html));
* In the paper itself, I added a citation to some recent related work in the GW astronomy field (parametric and nonparametric) so that a casual reader can have a look at what's going on;
* Concerning the mathematical details of the Dirichlet process, I added a couple of references where the details are given, both about the DP and the DPGMM. I feel that adding more details in that direction would make the paper too long with respect to the JOSS requirements (it was already almost hitting the 1000 words limit);
* The documentation is managed via readthedocs, and their workflow already updates the docs at every push.

I hope this addresses your comments!
