+++ figaro/rate.py
@@ -1,5 +1,6 @@
 import numpy as np
 import matplotlib.pyplot as plt
+import warnings
 from pathlib import Path
 from collections import Counter
 from scipy.stats import poisson, norm
