+++ README.md
@@ -5,7 +5,7 @@ It is also possible to perform hierarchical inferences: in this case, the model 
 Differently from other DPGMM implementations relying on variational algorithms, FIGARO does not require the user to specify a priori the maximum allowed number of mixture components. The required number of Gaussian distributions to be included in the mixture is inferred from the data. The documentation and user guide for FIGARO is available at [the documentation page](https://figaro.readthedocs.io).
 
 [![status](https://joss.theoj.org/papers/d5e757291a0287f2fef1466c7eacdff5/status.svg)](https://joss.theoj.org/papers/d5e757291a0287f2fef1466c7eacdff5)
-![Test](https://github.com/sterinaldi/FIGARO/actions/workflows/python-package.yml/badge.svg)
+![Test](https://github.com/sterinaldi/FIGARO/actions/workflows/test.yml/badge.svg)
 ## Getting started
 
 You can install FIGARO either via pip (stable release, recommended) 
