### 2024-05-20 6fdbbb3
check for components while saving
