### 2024-05-23 #6589 @Uddiptaatwork
Dear @sterinaldi @dfm apologies for the delay, we experienced a disruptive few weeks at the institute due to ongoing geopolitical issues.
I have just finished reviewing the paper draft and code documentation also keeping in mind the points raised by @dgerosa in this context.
I am currently running tests and reviewing installation specifics on MacOS(intel) and Linux and will confirm very shortly on the status.
