### 2024-05-23 #6589 @Uddiptaatwork
- Installation was seamless both from source and PyPI in Linux(deb) and MacOS(intel).
- The tests passed with 3 warnings however:
![Screenshot 2024-05-23 at 12 44 28](https://github.com/openjournals/joss-reviews/assets/50993054/0753ecf6-2abb-4420-9622-7d3f4ef78f4f)

All the warnings are however `DeprecationWarning`s and can be ignored for the time being. However, specific version updates to environment modules can cause future errors in executing certain parts of the code.
So, this is just a request to the authors to look into this just to ensure future compatibility.

In any case, I recommend this article for publication and congratulate the authors for a great piece of work.
