### 2024-05-23 #6589 @Uddiptaatwork
> Hi @Uddiptaatwork, thanks!
>
> Concerning the warnings, I'm not really sure about what's going on there, to be honest. I was aware of these warnings (MacOS, M1), but they are relative to a documentation line (the first two, the fact that I wrote `\int` in the documentation causes this warning) and a plot label (the math environment space in LaTeX). The code runs smoothly without warnings outside testing, so my best guess for this is something related to the testing suite. I'll put it on my to do list for future investigation anyway.
>
> Edit: I just checked, it's an issue with backslashes. I simply converted individual backslashes to double backslashes (https://stackoverflow.com/questions/52335970/how-to-fix-syntaxwarning-invalid-escape-sequence-in-python).

Perfect, thanks for looking into this!
@dfm I confirm my recommendation to publish this article.
