### 2024-05-25 48e1591
Tiny fixes for JOSS review
