+++ paper.bib
@@ -240,7 +240,7 @@ archivePrefix = {arXiv},
 @ARTICLE{edelman:2023,
        author = {{Edelman}, Bruce and {Farr}, Ben and {Doctor}, Zoheyr},
         title = "{Cover Your Basis: Comprehensive Data-driven Characterization of the Binary Black Hole Population}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Gravitational wave astronomy, Gravitational waves, Black holes, Compact objects, High energy astrophysics, 675, 678, 162, 288, 739, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Astrophysics of Galaxies},
          year = 2023,
         month = mar,
@@ -259,7 +259,7 @@ archivePrefix = {arXiv},
 @ARTICLE{toubiana:2023,
        author = {{Toubiana}, A. and {Katz}, Michael L. and {Gair}, Jonathan R.},
         title = "{Is there an excess of black holes around 20 M{\ensuremath{\odot}}? Optimizing the complexity of population models with the use of reversible jump MCMC.}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {black hole physics, gravitational waves, methods: statistical, General Relativity and Quantum Cosmology, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2023,
         month = oct,
