### 2024-05-25 #6589 @sterinaldi
Hi @dfm – I merged your PR. The updated version number of the code is v1.6.7 and the DOI is 10.5281/zenodo.11302325
